requirements.txt
fastapi~=0.115
uvicorn~=0.30
pydantic~=2.9
pydantic-settings~=2.4
jinja2~=3.1
streamlit~=1.39
dotenv
requests
google-genai
openai
pytest
ruff
mypy
black
aiohttp